+++ packages/hooks/src/index.js
@@ -5,3 +5,4 @@ export { default as useScrollingIntent } from './useScrollingIntent';
 export { default as useConnectModal } from './modal/useConnectModal';
 export { default as useOpenModal } from './modal/useOpenModal';
 export { default as useMutationFlow } from './useMutationFlow';
+export { default as useStorageState } from './useStorageState';