+++ packages/hooks/src/useStorageState/index.d.ts
@@ -1,0 +1,6 @@
+export interface Options {
+    key?: string;
+    defaultValue?: unknown;
+}
+
+export default function (opts: Options): [unknown, (value?: unknown) => void];